requirements.txt
flask
flask-cors
gunicorn
moviepy==1.0.3
numpy
pillow
rembg
imageio
imageio-ffmpeg